pyproject.toml
[project]
name = "scheduleit"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.5",
    "numpy>=2.1.3",
    "openai>=1.55.3",
    "pyaudio>=0.2.14",
    "python-dotenv>=1.0.1",
    "pyttsx3>=2.98",
    "speechrecognition>=3.11.0",
    "uvicorn>=0.32.1",
    "websockets>=14.1",
]

[dependency-groups]
dev = [
    "black>=24.10.0",
    "isort>=5.13.2",
]
